add sport to boostrary

+++ gfx/skia/skia/src/base/SkEndian.h
@@ -131,7 +131,7 @@ static inline void SkEndianSwap64s(uint64_t array[], int count) {
     #define SkTEndian_SwapLE64(n)    (n)
 #else   // SK_CPU_BENDIAN
     #define SkEndian_SwapBE16(n)    (n)
-    #define SkEndian_SwapBE32(n)    (n)
+    #define SkEndian_SwapBE32(n)    uint32_t(n)
     #define SkEndian_SwapBE64(n)    (n)
     #define SkEndian_SwapLE16(n)    SkEndianSwap16(n)
     #define SkEndian_SwapLE32(n)    SkEndianSwap32(n)
